refactor(providers): separation fetch and query providers
