+++ scripts/files/src/App/Routes/__tests__/Routes.test.tsx
@@ -1,4 +1,3 @@
-import { describe, it, expect } from 'vitest';
 import { PropsWithChildren } from 'react';
 import { render, screen, waitFor, act } from 'shared/testsUtils/customRender';
