+++ src/App/FetchProvider/index.ts
@@ -1,4 +1,3 @@
 export { default, FetchContext } from './FetchProvider';
-export type { FetchContextType } from './FetchProvider';
-export type { TResponse } from './FetchProvider';
-export { default as setResponseError } from './setResponseError';
+export type { FetchContextType, TFetchCustom } from './FetchProvider';
+export type { TResponse } from './FetchProvider.helpers';
