+++ src/App/NotificationProvider/Notifications/Notifications.scss
@@ -39,6 +39,7 @@ $typesLite: (
         background-color: inherit;
         min-width: inherit;
         padding: 0 0 0 1rem;
+        width: 2rem;
         &:after {
           display: none;
         }
