+++ src/App/QueryProvider/index.ts
@@ -1,0 +1,2 @@
+export { default } from './QueryProvider';
+export type { TQueryProvider } from './QueryProvider';
