--- src/shared/components/DownloadLink/__tests__/DownloadLink.test.tsx
@@ -1,20 +1,0 @@
-import { render } from '@testing-library/react';
-import { MODES } from 'shared/components/Loader';
-import DownloadLink from '../DownloadLink';
-
-const defaultProps = {
-  loaderMode: MODES.none,
-  onDownload: vi.fn(),
-};
-
-describe('DownloadLink', () => {
-  it('Render <DownloadLink/>', () => {
-    const { asFragment } = render(<DownloadLink {...defaultProps} />);
-    expect(asFragment()).toMatchSnapshot();
-  });
-
-  it('Render <DownloadLink/> disabled', () => {
-    const { asFragment } = render(<DownloadLink {...defaultProps} isDisabled />);
-    expect(asFragment()).toMatchSnapshot();
-  });
-});
