+++ src/shared/helpers/index.ts
@@ -1,7 +1,6 @@
 export { formatDate, isValidDate, setDate } from './formatDate';
 export { default as isEmptyOrNull, isEmptyArray, isEmptyObject } from './isEmptyOrNull';
 export { default as mergeObj } from './mergeObj';
-export { default as manageConfig } from './manageConfig';
 export { default as setAnomalyEmptyItems } from './setAnomalyEmptyItems';
 export { default as setConfirmClassModifier } from './setConfirmClassModifier';
 export { default as emptyFunction } from './emptyFunction';
