--- src/shared/testsUtils/WrapperQuery.tsx
@@ -1,16 +1,0 @@
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { ReactNode } from 'react';
-
-const WrapperQuery = ({ children, queryData = {} }: { children: ReactNode; queryData?: unknown }) => {
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        retry: false,
-        queryFn: () => queryData || {},
-      },
-    },
-  });
-  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
-};
-
-export default WrapperQuery;
