+++ src/shared/testsUtils/customRender.tsx
@@ -1,8 +1,9 @@
-import { render, RenderOptions } from '@testing-library/react';
+import { render, renderHook, RenderOptions } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
-import FetchProvider from 'App/FetchProvider';
-import UserProvider from 'App/UserProvider';
 import EnvironmentProvider from 'App/EnvironmentProvider';
+import UserProvider from 'App/UserProvider';
+import FetchProvider from 'App/FetchProvider';
+import QueryProvider from 'App/QueryProvider';
 import NotificationProvider from 'App/NotificationProvider';
 import { ReactElement, ReactNode } from 'react';
 import MOCK_API_URL from './constants';
@@ -18,6 +19,7 @@ const MockProviders =
     route = '/',
     oidcUser = { member_of: [`CN=${role}`], name: 'Samuel Gomez' },
     accessToken = 'accessToken',
+    queryData,
     ...testMock
   }: TMockProvider) =>
   ({ children }: { children: ReactNode }) => {
@@ -37,13 +39,20 @@ const MockProviders =
       },
     });
 
+    const queriesOptions = {
+      retry: false,
+      ...(queryData ? { queryFn: () => queryData } : {}),
+    };
+
     return (
       <EnvironmentProvider useEnvFn={useEnvFn}>
         <UserProvider useOidcUserFn={useOidcUserFn}>
           <FetchProvider useOidcAccessTokenFn={useOidcAccessTokenFn}>
-            <NotificationProvider>
-              <MemoryRouter initialEntries={[`${route}`]}>{children}</MemoryRouter>
-            </NotificationProvider>
+            <QueryProvider queriesOptions={queriesOptions}>
+              <NotificationProvider>
+                <MemoryRouter initialEntries={[`${route}`]}>{children}</MemoryRouter>
+              </NotificationProvider>
+            </QueryProvider>
           </FetchProvider>
         </UserProvider>
       </EnvironmentProvider>
@@ -53,10 +62,14 @@ const MockProviders =
 const customRender = (ui: ReactElement, options?: Omit<RenderOptions, 'wrapper'>, testMock: TMockProvider = { role: '' }) =>
   render(ui, { wrapper: MockProviders(testMock), ...options });
 
+type TcustomRenderHook = (testMock?: TMockProvider) => typeof renderHook;
+
+const customRenderHook: TcustomRenderHook = testMock => (hook, options) => renderHook(hook, { wrapper: MockProviders(testMock ?? {}), ...options });
+
 // re-export everything
 export * from '@testing-library/react';
 export { default as userEvent } from '@testing-library/user-event';
 export { axe } from 'jest-axe';
 
 // override render method
-export { customRender as render };
+export { customRender as render, customRenderHook };
