+++ src/shared/testsUtils/index.ts
@@ -2,7 +2,6 @@ export * from './customRender';
 export { default as emptyFunction } from '../helpers/emptyFunction';
 export { default as clearString } from './clearString';
 export { default as WrapperStaticRouter } from './WrapperStaticRouter';
-export { default as WrapperQuery } from './WrapperQuery';
 export { default as renderWithWrapperStaticRouter } from './renderWithWrapperStaticRouter';
 export { default as MOCK_API_URL } from './constants';
 export { default as renderWithContainer } from './renderWithContainer';
